feat: Rework shutting down for HTTP server to use a signal string
Revised `run_modes.go` to stop the HTTP server if the
"X-Server-Shutdown" header value matches the
`shutdownSignalHeaderValue`.
Now it's possible to shutdown the server based on header set in the request
rather than closing after a set number of responses. This provides more
efficiency and control over the server's operations.